[BBB-156] ✨Feat: 스터디 개설 및 가입 시 보증금 차감 로직 구현

+++ app/external-api/src/main/java/com/bombombom/devs/external/user/controller/dto/UserProfileResponse.java
@@ -11,7 +11,7 @@ public record UserProfileResponse(
     String image,
     String introduce,
     Integer reliability,
-    Integer money,
+    Long money,
     Role role
 ) {
