+++ domain/src/main/java/com/bombombom/devs/study/model/UserStudy.java
@@ -40,9 +40,9 @@ public class UserStudy extends BaseEntity {
     private Study study;
 
     @Column(name = "security_deposit")
-    private Integer securityDeposit;
+    private Long securityDeposit;
 
-    public static UserStudy of(User user, Study study, Integer securityDeposit) {
+    public static UserStudy of(User user, Study study, Long securityDeposit) {
         return UserStudy.builder()
             .user(user)
             .study(study)
